style: move create div for test fn and related stuff to main.ts
