--- src/components/directives/in-view/test/const.ts
@@ -1,12 +1,0 @@
-/*!
- * V4Fire Client Core
- * https://github.com/V4Fire/Client
- *
- * Released under the MIT license
- * https://github.com/V4Fire/Client/blob/master/LICENSE
- */
-
-/**
- * The top margin of test <div>, in pixels.
- */
-export const TEST_DIV_MARGIN_TOP_PX = 1000;
