+++ src/components/directives/in-view/test/helpers.ts
@@ -8,85 +8,16 @@
 
 import type { Locator, Page } from 'playwright';
 
-import type { Watcher, WatchHandler, WatchOptions } from 'core/dom/intersection-watcher';
-
-import { Component, Scroll } from 'tests/helpers';
-
-import { TEST_DIV_MARGIN_TOP_PX } from 'components/directives/in-view/test/const';
+import { Scroll } from 'tests/helpers';
 
 /**
- * A handler to pass to v-in-view
- * @param watcher - The parameter of the watch handler
+ * The top margin of the test `<div>`, in pixels
  */
-function handler(watcher: Watcher): void {
-	const div = watcher.target;
-
-	const previousValue = parseInt(
-		div.getAttribute('data-test-in-view') ?? '0',
-		10
-	);
-
-	const nextValue = previousValue + 1;
-	watcher.target.setAttribute('data-test-in-view', nextValue.toString());
-}
+export const TEST_DIV_MARGIN_TOP_PX = 1000;
 
 /**
- * Returns the value of the watcher call counter stored in given locator.
- * @param locator - The source locator
- */
-export async function getWatcherCallsCount(locator: Locator): Promise<number | null> {
-	const storedValue = await locator.getAttribute('data-test-in-view');
-
-	if (storedValue == null) {
-		return null;
-
-	}
-
-	return parseInt(storedValue, 10);
-}
-
-/**
- * Force adds test handler (which counts calls) to given watch,
- * or replaces given one if a function/undefined is provided.
- *
- * @param watch - A watch parameter to update
- */
-function addTestHandlerToWatch(
-	watch: CanUndef<WatchHandler | Partial<WatchOptions>>
-): WatchHandler | WatchOptions & { handler: WatchHandler } {
-
-	if (Object.isUndef(watch) || Object.isFunction(watch)) {
-		return handler;
-
-	}
-
-	return {...watch, handler};
-}
-
-/**
- * Creates a <div> element with v-in-view set to inViewValue.
- *
- * @param page - The target page.
- * @param inViewValue - The value of v-in-view directive.
- */
-export async function createDivForInViewTest(
-	page: Page, inViewValue: CanUndef<CanArray<WatchHandler | Partial<WatchOptions>>>
-): Promise<Locator> {
-
-	await Component.createComponent(page, 'div', {
-		'v-in-view': Object.isArray(inViewValue) ?
-			inViewValue.map(addTestHandlerToWatch) :
-			addTestHandlerToWatch(inViewValue),
-		'data-testid': 'div',
-		style: `margin-top: ${TEST_DIV_MARGIN_TOP_PX}px; width: 20px; height: 20px`
-	});
-
-	return page.getByTestId('div');
-}
-
-/**
- * Initializes the viewport by making test <div> be not in it when the test is started.
- * @param page - The target page.
+ * Initializes the viewport by making test <div> be not in it when the test is started
+ * @param page - the target page
  */
 export async function initViewport(page: Page): Promise<void> {
 	await page.setViewportSize({width: 500, height: TEST_DIV_MARGIN_TOP_PX / 2});
@@ -99,9 +30,9 @@ export async function initViewport(page: Page): Promise<void> {
 }
 
 /**
- * Makes element pointed by locator enter the viewport.
+ * Makes element pointed by locator enter the viewport
  *
- * @param locator - The target locator
+ * @param locator - the target locator
  * @see https://playwright.dev/docs/api/class-locator#locator-click
  */
 export async function makeEnterViewport(locator: Locator): Promise<void> {
@@ -109,8 +40,8 @@ export async function makeEnterViewport(locator: Locator): Promise<void> {
 }
 
 /**
- * Restores viewport set in initViewport().
- * @param page - The target page
+ * Restores viewport set in initViewport()
+ * @param page - the target page
  */
 export async function restoreViewport(page: Page): Promise<void> {
 	await Scroll.scrollToTop(page);
